execute: fix argument counting regression from 0ae4f887
operator+ binds more tightly than ?:, so writing

	a + b ? 5 : 4

is equivalent to:

	((a + b) != 0) ? 5 : 4

The surrounding code intends to copy 'a' elements, then possibly one
more if b is not zero.  The expression in 0ae4f887 only works correctly
when 'a' is small.  However, there is no specific bound on 'a', so for
large 'a', this produces an allocation which is too small.  The
overwrite is mostly concealed by the quick use of execvp.  Very large
argument lists will produce a crash with this code.  For example,

	lxc-execute -n test -- /usr/bin/printf '%s\n' {1..100000}

Change the code to the intended behavior of:

	a + 3 + (b ? 1 : 0)

Use 3 instead of 4 since there is no apparent need for a fourth element.