pivot: move initial chdir above pivot
Previously, the current working directory was set to the root of the new
rootfs if a pivot was used and was left unchanged if no pivot was used.
Now, it is always set to the root directory of the container, which will
be the root of the host if no separate rootfs is prepared.