Fix buffer overflow in socket name setup
Switch to snprintf to detect and reject excessive length names.  Without
this, a container with a name longer than UNIX_PATH_MAX will smash a
stack buffer with user chosen bytes.

This is not exploitable in the normal configuration since none of the
tools are setuid.  It may be exploitable on systems where the tools have
been made setuid, or where a setuid wrapper runs the tools on behalf of
an otherwise unprivileged user.